pyproject.toml
[tool.poetry]
name = "streamlit-connection"
version = "0.1.0"
description = "An experimental project to standardize streamlit connections"
authors = ["Zachary Blackwood <zachary@streamlit.io>"]
readme = "README.md"
packages = [{include = "streamlit_connection", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<3.11"
streamlit = "^1.13.0"
pymongo = "^4.2.0"
snowflake-connector-python = "2.8.0"
asn1crypto = "1.5.1"
certifi = "2022.9.14"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
cryptography = "36.0.2"
filelock = "3.8.0"
idna = "3.4"
oscrypto = "1.3.0"
pycparser = "2.21"
pycryptodomex = "3.15.0"
PyJWT = "2.5.0"
pyOpenSSL = "22.0.0"
pytz = "2022.2.1"
requests = "2.28.1"
typing-extensions = "4.3.0"
urllib3 = "1.26.12"
pyarrow = "^9.0.0"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.3.5"}

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.9.23"
pytest = "^7.1.3"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
skip = ["./.venv", "./direnv", ".env"]

[tool.black]
exclude = '''
(
  /(
      \.vscode
    | \.git
    | \.pytest_cache
    | \.mypy_cache
    | \.venv
    | \.env
    | \.direnv
  )/
)
'''
include = '\.pyi?$'
line-length = 88

[tool.mypy]
files = [
  "**/*.py",
]
follow_imports = "silent"
ignore_missing_imports = true
scripts_are_modules = true